pyproject.toml
[project]
name = "asimov-academy-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=2.3.2",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.122.0",
    "groq>=0.36.0",
    "ipykernel>=7.1.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.13",
    "yfinance>=0.2.66",
]
